Dropping commented-out bits in app.go; leaving only type alias
